% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano.R
\name{nano}
\alias{nano}
\title{Generate Nano IDs}
\usage{
nano(
  n = 1,
  length = 16,
  prefix = "",
  characters = c(base::LETTERS, base::letters, 0:9)
)
}
\arguments{
\item{n}{Integer scalar. How many IDs to generate.
Must be a positive whole number. Default is \code{1}.}

\item{length}{Integer scalar. Total length \strong{including} any \code{prefix}.
Default is \code{16}.}

\item{prefix}{Single string (possibly empty) to prepend to every ID.
Must be shorter than \code{length}.}

\item{characters}{Character vector from which the random part is drawn.
Defaults to \code{c(LETTERS, letters, 0:9)}.}
}
\value{
A character vector of length \code{n} containing the generated IDs.
}
\description{
Quickly create random, URL-safe identifier strings ("nano IDs").
The generator is \strong{not} cryptographically secure, but it is fast and
compact—ideal for temporary keys, file names, etc.
}
\examples{
nano()
nano(24)
nano(prefix = "Q")
nano(5, prefix = "Q", characters = 0:9)
nano(n = 10)

}
