% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple.R
\name{triple}
\alias{triple}
\title{Construct an RDF Triple Statement}
\usage{
triple(subject, predicate, object)
}
\arguments{
\item{subject}{RDF subject (URI or blank node).}

\item{predicate}{RDF predicate (URI).}

\item{object}{RDF object (URI, literal, or blank node).}
}
\value{
A formatted RDF triple as a string.
}
\description{
This helper constructs an RDF triple statement as a formatted string, combining a subject, predicate, and object.
}
\examples{
triple(uri("https://example.com/1234"), uri("https://example.com/hasName"), literal("Apple"))
triple(uri("1234", prefix = "https://example.com/"), uri("rdf:type"), uri("https://example.com/Fruit"))

}
