% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed.R
\name{typed}
\alias{typed}
\title{Construct a Typed RDF Literal}
\usage{
typed(x, datatype = NULL)
}
\arguments{
\item{x}{Character or numeric value. The content to be typed.}

\item{datatype}{Character string. The XML Schema datatype (e.g., \code{"string"}, \code{"integer"}). Must be provided.}
}
\value{
A formatted RDF typed literal.
}
\description{
This helper constructs a typed RDF literal, associating a string value with a specific XML Schema datatype (\code{xsd}).
}
\details{
Supported datatypes are: \code{"string"}, \code{"integer"}, \code{"decimal"}, \code{"boolean"}, \code{"date"}, \code{"dateTime"}.
}
\examples{
typed("42", "integer")
typed("true", "boolean")
typed("2025-03-30", "date")

}
