% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langstring.R
\name{langstring}
\alias{langstring}
\title{Construct a Language-tagged RDF literal}
\usage{
langstring(x, lang = NULL)
}
\arguments{
\item{x}{Character string. The textual content of the literal.}

\item{lang}{Character string. The language tag conforming to \href{https://tools.ietf.org/html/bcp47}{BCP47}. Must be provided.}
}
\value{
A formatted RDF language-tagged literal (\code{rdf:langString}).
}
\description{
This helper constructs an RDF language-tagged literal (\code{rdf:langString}), which is a string explicitly associated with a language tag following BCP47 conventions (e.g., \code{"Apple"@en}, \code{"Apfel"@de-CH}).
}
\examples{
langstring("Apple", "en")
langstring("Grüezi", "de-CH")

}
