% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed.R
\name{typed}
\alias{typed}
\title{Construct a Typed RDF Literal}
\usage{
typed(x, datatype = NULL)
}
\arguments{
\item{x}{Character or numeric value. The content to be typed.}

\item{datatype}{Character string. The XML Schema datatype (e.g., \code{"string"}, \code{"integer"}). Must be provided.}
}
\value{
A formatted RDF typed literal. If \code{x} is any of \code{NA}, \code{NULL}, or \code{""}, the return value will be \code{NULL}.
(To avoid things like \verb{"NA"^^<http://www.w3.org/2001/XMLSchema#integer>} as a result of \code{typed(NA, "integer")})
}
\description{
This helper constructs a typed RDF literal, associating a string value with a specific XML Schema datatype (\code{xsd}).
}
\details{
Supported datatypes are: \code{"string"}, \code{"integer"}, \code{"decimal"}, \code{"boolean"}, \code{"date"}, \code{"dateTime"}.
}
\examples{
typed("42", "integer")
typed("true", "boolean")
typed("2025-03-30", "date")
typed("", "integer") # this will return `NULL`

}
